Refactor: rename `is-server` to `platform`
