+++ libs/core/src/utils/helpers/platform.ts
@@ -3,3 +3,8 @@
  */
 export const isServer =
   typeof process !== 'undefined' && process.env !== undefined
+
+/**
+ * Simple UA check to determine if the platform is iOS
+ */
+export const isIOS = /iPad|iPhone|iPod/.test(navigator.userAgent)
